Update the folder structure and build tts modules
The app now follows a generic folder structure with
the `App` folder being the main application definition.
The modules are located inside App/Modules.

As for the tts, `edge_tts` package is used (Modules/voice_tts)
to create functions that input a str and play a temporarily
generated mp3

This commit is a general commit and is aimed to provide scope
for initial testing. The `App/main.py` is bound to change
in future commits.